Fix out-of-order gzip headers (#80)
We were spending some time just moving 10 bytes from the end of a file
to the start of the next file. This was only necessary because we called
gzip.Reader.Reset() before we created the Next() file. Reset() reads the
next gzip.Header, which was getting teed into the previous file instead
of the next file.

The fix for this made unsafe assumptions about the length of a gzip
header, which frightened me, but this change makes that go away. This
change also makes a lot of wasted work go away (essentially copying the
entire data section for each APK twice).

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>