Add global expanded APK cache (#133)
This dedupes fetching and expansion of APKs to avoid simultaneous builds
doing a lot of duplicate expensive work. As a bonus, we also deduplicate
indexing an APK's tar headers so we do it once per APK per process
instead of every time we open an APK.

The caching mechanisms that already existed are very useful for cache
hits on subsequent builds, but there was still some overhead loading
things from disk per cache hit. We avoid most of that now. There was
also no coordination across builds when fetching indexes, keys, and
APKs. Now, we should only fetch or load any given thing one time, which
will reduce the number of requests we send out to the internet but also
a lot of CPU in gunzip and untar.

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>