Sort packages with bad versions correctly (#146)
Previously, if sortPackage's less function encountered an error while
parsing the version of a package, it would always return false. I can
only assume this is a mistake that stems from returning a zero value
when encountering an error, but in this case that orders packages
improperly because we'd return false (asserting j > i) if either version
i or j was invalid, which will not produce good output.

When sorting, we now consider versions that can be parsed as a closer
match than versions that can't be parsed. So if version[i] can parse but
version[j] can't parse, version[j] will come after version[i] when we
sort them (all else being equal).

Signed-off-by: Jon Johnson <jon.johnson@chainguard.dev>